Started refactoring the Map functionality
The first step is to introduce more restrictive overloads to HasMany that
have different return types based on their intent. Element maps are the
first target, with overloads limited to value types.

Created a MapElementBuilder, which ditches the baggage of ToManyBase and
starts anew. So far creates the <map> and <element> correctly, still
gotta support the <key> and <index>.