Implemented a value-type/entity map builder
Refactored the HasMany methods to include value type key'd dictionaries
with values of entities. These are restricted in similar ways to the
element maps by explicitly specifying the value type methods as
overloads.

These methods default to many-to-many collections even though they're
created through HasMany. This is a bit confusing, but one-to-many really
doesn't make much sense to me in this situation; it creates a schema
that I can't see ever being useful. If this causes confusion we should
change it...