Added composite-element support to new builders
Renamed MapManyToManyBuilder to MapValueKeyBuilder as it's hard to
distinguish between entity maps and composite-element maps (they're both
classes). The renamed builder now has both routes on, making it a bit of
a general case for anything with a value type as a key. This class might
end up being even more generic, but we'll see (I'd like to try to avoid
that if possible).