Unified the builders into a single MapBuilder
The default implementations (element, entity map, etc...) are decided
programatically instead of using generic constraints. This actually
gives us more power because we can do some proper checks against the
types, instead on relying on the rather lackluster constraint system
(for example: we can detect enums as element mappings).

Also updated the majority of the tests that were relying on the AsMap
method to initiate the <map>. This is a breaking change. The AsMap
method is now not available for maps, and most of the time all the
values should use reasonable defaults.